How did I manage this?
1) Finished Step 5 of Part I in the Plan of Action. There is now a
consistent system in place to obtain, line-by-line, the details
of each operation. This includes the type of the operation, the
virtual registers and their order of appearance, and the constant that
is present (if applicable).
2) Added a function, nextNum(), which is integral to obtaining
information about an ILOC instruction. Because ILOC instructions have
specific formats depending on their types, we can assume that
parsing the numeric values in order can provide us the operations
in order.

Step 5 was tested with block1.i from the test blocks, which contains
a good portion of the operations. The logic is pretty much the same
between each operation type, so this should be acceptable for
this moment. It's too tedious to read 60 ILOC instructions and compare
them with 60*6 lines of output.