Finished the groundwork for live-focused top-down allocation, and also refactored simple top-down a bit.
1) Refactored simple top-down allocation again to ignore r0 when
considering the number of virtual registers (length of sorted array).
2) Wrapped up changes in live-focused top-down allocation, and prepared
skeleton for the function.
3) Created the algorithm for live-focused top-down allocation,
in methods to implement.txt.

I haven't debugged the functions to be used in live-focused top-down
allocation, but it's doable if I chip at it bit by bit. If all goes well and the
design is valid, I should be able to finish live top-down allocation and
just call opSimpleTd() for output. Hopefully I can refactor the intNode
deletions into a function soon to save real estate.

Personal note:
I'm not in the best mental shape right now due to a family loss that
occurred today, but I'll push on and get this done. It's due Friday, so
I'll manage it by Wednesday or Thursday even with Tuesday's midterm. :)
I suspect the report will be easy-peasy with these details sorted out,
and it's due Sunday anyways (I've written plenty of very good 15-page
documentations in a day, so 6 pages is a cinch).