I basically tore down my lecture top-down allocation code, so I can rewrite it from scratch (though I did keep the idea of sorting the registers in ascending order of occurrence, to make determining eviction easier).
1) Removed trackLiveRegs()
2) Removed non-skeleton code from topDownLive()
3) Removed function macro for trackLiveRegs()
4) Updated my "Methods to Implement" document accordingly.

As seen in previous commits, I tend to do well-organized and refactored work when I'm
focused, but most of the code for the current stage was written while I
was distressed and honestly should have been taking more time out. I decided it'd be best to take what I've gathered from this cycle and start fresh.